Moving getComposedData into one separately file
